disk: Use a checkbox tree viewer for the disk IO view
Reviewed-on: https://git.eclipse.org/r/110349
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>

+++ analysis/org.eclipse.tracecompass.analysis.os.linux.core.tests/stubs/org/eclipse/tracecompass/analysis/os/linux/core/tests/stubs/inputoutput/IoTestCase.java
@@ -81,7 +81,7 @@ public class IoTestCase extends LinuxTestCase {
          *
          * @return The expected number of sectors
          */
-        public long getNbSectors() {
+        public double getNbSectors() {
             return fSectors;
         }
     }
